Refactoring: CurveUpdateUtils => StratUpdateUtils
